% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberize.R
\name{countReplicates}
\alias{countReplicates}
\title{Count the replicates in the column}
\usage{
countReplicates(data, colName)
}
\arguments{
\item{data}{the matrix data that will be used.}

\item{colName}{Column name that will be counted the replicates.}
}
\description{
Count the replicates in the column
}
