% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numberize.R
\name{NumberizeNames}
\alias{NumberizeNames}
\title{Numberize Names}
\usage{
NumberizeNames(data, colName)
}
\arguments{
\item{data}{Data is the input matrix.}

\item{colName}{ColName is the column name to specify which column will be numberized.}
}
\description{
Numberize Names
}
